+++ frontend/package.json
@@ -29,6 +29,7 @@
     "bootstrap": "^5.3.3",
     "eslint-config-airbnb-base": "^15.0.0",
     "formik": "^2.4.6",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.2",
     "react-dom": "^18.2.0",
