+++ frontend/src/Commons/Forms/InputModules/FieldsAndSoil/FieldsAndSoil.style.ts
@@ -11,10 +11,6 @@ import styled from '@emotion/styled';
 import screenSizes from '@Constants/ScreenSize';
 import * as tokens from '@bcgov/design-tokens/js';
 
-type StyledListType = {
-  width?: string;
-};
-
 const StyledFarmInfo = styled.div`
   display: flex;
   flex-direction: column;
@@ -60,90 +56,6 @@ const StyledButtonGroupContainer = styled.div`
     width: 100%;
   }
 `;
-
-const StyledListContainer = styled.div`
-  position: relative;
-  display: flex;
-  flex-direction: row;
-  padding-top: 20px;
-  gap: 20px;
-  @media (min-width: ${screenSizes.desktop}) {
-  }
-`;
-
-const StyledListItem = styled.div<StyledListType>`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  width: ${(props) => props.width};
-
-  h4 {
-    font: ${tokens.typographyBoldBody};
-  }
-  @media (min-width: ${screenSizes.desktop}) {
-    h4 {
-      font: ${tokens.typographyBoldH6};
-    }
-  }
-`;
-
-const StyledFontAwesomeContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  font-size: 24spx;
-  gap: 30px;
-  margin-right: 10px;
-  padding-top: 8px;
-
-  @media (min-width: ${screenSizes.desktop}) {
-    font-size: 32px;
-    padding-top: 20px;
-  }
-`;
-
-const StyledFieldInfoList = styled.div`
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  @media (min-width: ${screenSizes.desktop}) {
-  }
-`;
-
-const StyledCommentContainerDesktop = styled.div`
-  display: none;
-  @media (min-width: ${screenSizes.desktop}) {
-    display: flex;
-    flex-direction: column;
-  }
-`;
-
-const StyledCommentContainerMobile = styled.div`
-  display: flex;
-
-  @media (min-width: ${screenSizes.desktop}) {
-    display: none;
-  }
-`;
-
-const StyledDivider = styled.div`
-  font-size: 18px;
-  font-weight: 400;
-  line-height: 30.61px;
-  display: flex;
-  align-items: center;
-  text-align: center;
-  color: ${tokens.typographyColorPlaceholder};
-  flex: 1;
-  content: '';
-  padding: 1px;
-  background-color: ${tokens.typographyColorPlaceholder};
-  width: 100%;
-  margin: auto;
-
-  @media (min-width: ${screenSizes.desktop}) {
-  }
-`;
-
 const StyledButtonContainer = styled.div`
   margin-top: 20px;
   width: 50%;
@@ -153,7 +65,7 @@ const StyledButtonContainer = styled.div`
 const StyledAddCancelButtonContainer = styled.div`
   display: flex;
   position: absolute;
-  right: 0;
+  right: 17%;
   gap: 10px;
 `;
 const StyledNewFieldButtonContainer = styled.div`
@@ -166,13 +78,6 @@ export {
   StyledFarmInfo,
   StyledTextAreaContainer,
   StyledButtonGroupContainer,
-  StyledListContainer,
-  StyledListItem,
-  StyledFontAwesomeContainer,
-  StyledFieldInfoList,
-  StyledCommentContainerDesktop,
-  StyledCommentContainerMobile,
-  StyledDivider,
   StyledButtonContainer,
   StyledAddCancelButtonContainer,
   StyledNewFieldButtonContainer,
